+++ modules/features/player/src/main/java/au/com/shiftyjelly/pocketcasts/player/view/ChapterProgressBar.kt
@@ -8,6 +8,7 @@ import android.graphics.Path
 import android.graphics.RectF
 import android.util.AttributeSet
 import android.view.View
+import au.com.shiftyjelly.pocketcasts.ui.helper.colorIntWithAlpha
 import au.com.shiftyjelly.pocketcasts.ui.theme.Theme
 import au.com.shiftyjelly.pocketcasts.ui.theme.ThemeColor
 import au.com.shiftyjelly.pocketcasts.utils.extensions.dpToPx
@@ -33,8 +34,16 @@ class ChapterProgressBar @JvmOverloads constructor(
     var theme: Theme.ThemeType = Theme.ThemeType.DARK
         set(value) {
             field = value
-            backgroundPaint.color = ThemeColor.playerContrast06(field)
-            progressPaint.color = ThemeColor.playerContrast06(field)
+            backgroundPaint.color = when {
+                field.darkTheme -> ThemeColor.playerContrast06(field)
+                field == Theme.ThemeType.LIGHT_CONTRAST -> ThemeColor.primaryUi05(field)
+                else -> ThemeColor.primaryUi02(field)
+            }
+            progressPaint.color = when {
+                field == Theme.ThemeType.ROSE -> ThemeColor.primaryIcon02Selected(field).colorIntWithAlpha(25)
+                field.darkTheme -> ThemeColor.playerContrast06(field).colorIntWithAlpha(25)
+                else -> Color.BLACK.colorIntWithAlpha(25)
+            }
         }
 
     val backgroundPaint = Paint().apply {
@@ -42,7 +51,7 @@ class ChapterProgressBar @JvmOverloads constructor(
         this.style = Paint.Style.FILL
     }
 
-    val progressPaint = Paint().apply {
+    var progressPaint = Paint().apply {
         this.color = Color.argb(50, 255, 255, 255)
         this.style = Paint.Style.FILL
     }