+++ modules/features/shared/src/test/java/au/com/shiftyjelly/pocketcasts/shared/AppLifecycleObserverTest.kt
@@ -34,6 +34,7 @@ class AppLifecycleObserverTest {
     @Mock @ApplicationContext private lateinit var context: Context
     @Mock private lateinit var settings: Settings
     @Mock private lateinit var autoPlayNextEpisodeSetting: UserSetting<Boolean>
+    @Mock private lateinit var useUpNextDarkThemeSetting: UserSetting<Boolean>
     @Mock private lateinit var packageUtil: PackageUtil
     @Mock private lateinit var appLifecycleAnalytics: AppLifecycleAnalytics
     @Mock private lateinit var preferencesFeatureProvider: PreferencesFeatureProvider
@@ -47,6 +48,7 @@ class AppLifecycleObserverTest {
     @Before
     fun setUp() {
         whenever(settings.autoPlayNextEpisodeOnEmpty).thenReturn(autoPlayNextEpisodeSetting)
+        whenever(settings.useDarkUpNextTheme).thenReturn(useUpNextDarkThemeSetting)
 
         whenever(appLifecycleOwner.lifecycle).thenReturn(appLifecycle)
 
@@ -75,6 +77,7 @@ class AppLifecycleObserverTest {
 
         verify(appLifecycleAnalytics).onNewApplicationInstall()
         verify(autoPlayNextEpisodeSetting).set(true)
+        verify(useUpNextDarkThemeSetting).set(false)
 
         verify(appLifecycleAnalytics, never()).onApplicationUpgrade(any())
     }
@@ -91,6 +94,7 @@ class AppLifecycleObserverTest {
         verify(appLifecycleAnalytics).onNewApplicationInstall()
 
         verify(autoPlayNextEpisodeSetting, never()).set(any(), any(), any())
+        verify(useUpNextDarkThemeSetting).set(false)
         verify(appLifecycleAnalytics, never()).onApplicationUpgrade(any())
     }
 
@@ -106,6 +110,7 @@ class AppLifecycleObserverTest {
         verify(appLifecycleAnalytics).onNewApplicationInstall()
 
         verify(autoPlayNextEpisodeSetting, never()).set(any(), any(), any())
+        verify(useUpNextDarkThemeSetting).set(false)
         verify(appLifecycleAnalytics, never()).onApplicationUpgrade(any())
     }
 
@@ -122,5 +127,6 @@ class AppLifecycleObserverTest {
 
         verify(appLifecycleAnalytics, never()).onNewApplicationInstall()
         verify(autoPlayNextEpisodeSetting, never()).set(any(), any(), any())
+        verify(useUpNextDarkThemeSetting, never()).set(any(), any(), any())
     }
 }